requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pydantic==2.11.1
pydantic_core==2.33.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-mock==1.12.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
tabula-py==2.10.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
